+++ src/refactoring/features/cart/ui/index.ts
@@ -1,0 +1,4 @@
+export { ProductItem } from './ProductItem';
+export { CartItemList } from './CartItemList';
+export { CouponSection } from './CouponSection';
+export { OrderSummary } from './OrderSummary';
